fix return_menu and refactor create_dish_menu
