--- test.txt
@@ -1,2 +1,0 @@
-HELLO MISHA and MISHA
-HELLO MISHA and Danil
